Fixed subcompany code behaviour : keep the value even when there's no custom VA
