+++ corekit/src/main/java/com/midtrans/sdk/corekit/models/BcaBankTransferRequestModel.java
@@ -11,6 +11,12 @@ public class BcaBankTransferRequestModel extends BankTransferRequestModel {
 
     @SerializedName("free_text")
     private FreeText freeText;
+    @SerializedName("sub_company_code")
+    private String subCompanyCode;
+
+    public BcaBankTransferRequestModel() {
+        super();
+    }
 
     public BcaBankTransferRequestModel(String vaNumber) {
         super(vaNumber);
@@ -19,5 +25,13 @@ public class BcaBankTransferRequestModel extends BankTransferRequestModel {
     public BcaBankTransferRequestModel(String vaNumber, FreeText freeText) {
         super(vaNumber);
         this.freeText = freeText;
+    }
+
+    public String getSubCompanyCode() {
+        return subCompanyCode;
+    }
+
+    public void setSubCompanyCode(String subCompanyCode) {
+        this.subCompanyCode = subCompanyCode;
     }
 }
