+++ corekit/src/main/java/com/midtrans/sdk/corekit/models/snap/BankTransferRequestModel.java
@@ -1,6 +1,5 @@
 package com.midtrans.sdk.corekit.models.snap;
 
-import android.text.TextUtils;
 import com.google.gson.annotations.SerializedName;
 
 import java.io.Serializable;
@@ -12,8 +11,8 @@ import java.io.Serializable;
 public class BankTransferRequestModel implements Serializable {
     @SerializedName("va_number")
     private String vaNumber;
-    @SerializedName("sub_company_code")
-    private String subCompanyCode;
+
+    public BankTransferRequestModel() {}
 
     public BankTransferRequestModel(String vaNumber) {
         setVaNumber(vaNumber);
@@ -25,15 +24,5 @@ public class BankTransferRequestModel implements Serializable {
 
     public void setVaNumber(String vaNumber) {
         this.vaNumber = vaNumber;
-    }
-
-    public String getSubCompanyCode() {
-        return subCompanyCode;
-    }
-
-    public void setSubCompanyCode(String subCompanyCode) {
-        if (!TextUtils.isEmpty(vaNumber)) {
-            this.subCompanyCode = subCompanyCode;
-        }
     }
 }
